refactor: update namespaces for serialization classes
